Move graph file creation to separate module
